@@ -1,24 +1,23 @@
 {
-  "name": "wallet-demo",
-  "version": "0.1.0",
+  "name": "shipl-demo",
+  "version": "1.0.0",
   "private": true,
   "homepage": "https://demo.shipl.co",
   "dependencies": {
-    "axios": "^0.18.0",
-    "ethers": "^4.0.27",
-    "nes.css": "^2.1.1",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
+    "nes.css": "^2.3.0",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
     "react-ga": "^2.5.7",
-    "react-scripts": "3.0.0",
-    "shipl-wallet": "^0.4.0",
+    "react-scripts": "^3.4.1",
+    "react-typist": "^2.0.5",
+    "shipl": "^2.0.0",
     "styled-components": "^4.2.0",
-    "web3": "^1.0.0-beta.54"
+    "web3": "^1.2.6"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test",
+    "test": "node test/puppeteer",
     "eject": "react-scripts eject",
     "deploy": "npm run build && surge build poke-shipl.surge.sh"
   },
@@ -36,5 +35,13 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "body-parser": "^1.19.0",
+    "express": "^4.17.1",
+    "puppeteer": "^2.1.1",
+    "request": "^2.88.2",
+    "request-promise": "^4.2.5",
+    "webpack-cli": "^3.3.11"
   }
 }
