@@ -1,185 +1,42 @@
 import React from 'react'
 import ReactGA from 'react-ga'
+import Shipl from 'shipl'
+import Web3 from 'web3'
 import Card from './Card'
 import Pokedex from './Pokedex'
-import styled from 'styled-components'
-import Web3 from 'web3'
-import Shiplwallet from 'shipl-wallet'
+import TextBubble from './TextBubble'
+import CardEvolution from './CardEvolution'
+import contractAbi from './contractAbi'
+import StyledApp from './style'
 
-const appId = '5fc6c72e-db33-4829-9a81-a4227b96238c'
-const targetContractAddress = '0x1c314cc1f12c6ae930385de21024cf32b63ffa0d'
+import CharmanderEvolution from './img/charmander-evolution.png'
+import BalbasaurEvolution from './img/balbasaur-evolution.png'
+import SquirtleEvolution from './img/squirtle-evolution.png'
 
-const contractAbi = [
-  {
-    constant: true,
-    inputs: [
-      {
-        name: '',
-        type: 'address'
-      }
-    ],
-    name: 'owners',
-    outputs: [
-      {
-        name: '',
-        type: 'uint256'
-      }
-    ],
-    payable: false,
-    stateMutability: 'view',
-    type: 'function'
-  },
-  {
-    constant: false,
-    inputs: [
-      {
-        name: 'pokemon',
-        type: 'uint256'
-      }
-    ],
-    name: 'claim',
-    outputs: [],
-    payable: false,
-    stateMutability: 'nonpayable',
-    type: 'function'
-  }
-]
-
-const StyledApp = styled.div`
-  font-family: "Press Start 2P";
-  font-weight: 400;
-  line-height: 1.5;
-  color: #212529;
-  text-align: left;
-  margin: 20px;
-
-  header {
-    position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    z-index: 9;
-    border-bottom: 4px solid #d3d3d3;
-    background-color: white;
-  }
-
-  header > .container {
-    display: flex;
-    align-items: baseline;
-    max-width: 1080px;
-    margin: 0 auto;
-    padding-top: 1rem;
-    transition: all 0.2s ease;
-  }
-
-  header > .container > .nav-brand {
-    margin-right: auto;
-  }
-
-  .nav-brand .brand-logo {
-    margin-right: 1rem;
-  }
-
-  .container {
-    max-width: 980px;
-    margin: 0 auto;
-    margin-top: 150px;
-  }
-
-  .card-board {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-  }
-
-  .showcase {
-    margin-top: 2.5rem;
-  }
-
-  section.message-list {
-    display: flex;
-    flex-direction: column;
-  }
-  .message-list > .message.-left {
-    align-self: flex-start;
-  }
-
-  .message-list > .message.-right {
-    align-self: flex-end;
-  }
-
-  .message-list > .message {
-    display: flex;
-    margin-top: 2rem;
-    align-items: flex-end;
-    align-items: end;
-  }
-
-  img {
-    margin: 10px;
-    max-width: 96px;
-  }
-
-  .half {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    flex-wrap: wrap;
-  }
-
-  footer {
-    margin-top: 30px;
-    text-align: center;
-  }
-
-  @media (max-width: 380px) {
-    .nes-prof,
-    .nes-bulbasaur,
-    .nes-charmander,
-    .nes-squirtle,
-    .nes-pokeball {
-      transform: scale(0.8);
-    }
-
-    p {
-      line-height: 1.2;
-      font-size: 12px;
-    }
-
-    h1 {
-      font-size: 1.5em;
-    }
-
-    .nav-brand {
-      margin: 0px 10px 0px 10px !important;
-    }
-
-    .nes-container {
-      padding: 1rem 1.2rem;
-    }
-
-    .pokedex-info {
-      padding-left: 10px;
-      margin-left: 10px;
-    }
-  }
-`
+const appId = 'kovan-821c17fe-df90-4250-a61f-3ea6f59d5f59'
+const targetContractAddress = '0x87Dc39A69D0E86840632cFae9Fc9d14325Aa39C2'
 
 const cards = [
   {
     id: 1,
     name: 'Bulbasaur',
-    picture: 'nes-bulbasaur'
+    picture: 'nes-bulbasaur',
+    pictureEvolution: BalbasaurEvolution,
+    idEvolution: 11
   },
   {
     id: 2,
     name: 'Charmander',
-    picture: 'nes-charmander'
+    picture: 'nes-charmander',
+    pictureEvolution: CharmanderEvolution,
+    idEvolution: 22
   },
   {
     id: 3,
     name: 'Squirtle',
-    picture: 'nes-squirtle'
+    picture: 'nes-squirtle',
+    pictureEvolution: SquirtleEvolution,
+    idEvolution: 33
   }
 ]
 
@@ -187,6 +44,8 @@ class App extends React.Component {
   constructor (props) {
     super(props)
     this.state = {
+      ownAPokemon: false,
+      ownAEvolution: false,
       poke: undefined,
       isLoading: true,
       txHash: undefined,
@@ -195,39 +54,40 @@ class App extends React.Component {
       catch: undefined,
       shiplwallet: undefined,
       web3: undefined,
+      eth: undefined,
       contract: undefined
     }
 
     this.getAccount = this.getAccount.bind(this)
     this.claimPokemon = this.claimPokemon.bind(this)
+    this.claimEvolution = this.claimEvolution.bind(this)
     this.pokedex = this.pokedex.bind(this)
   }
 
-  async componentWillMount () {
-    const shiplwallet = await Shiplwallet.create({ appId })
+  async componentDidMount () {
+    const shiplwallet = await Shipl.create({ appId, provider: 'shiplwallet' })
     const web3 = new Web3(shiplwallet.getWeb3Provider())
     const contract = new web3.eth.Contract(contractAbi, targetContractAddress)
     this.setState({
       web3,
       contract,
       shiplwallet,
-      network: shiplwallet.shiplID.network,
-      identity: shiplwallet.shiplID.auth.identity
+      network: shiplwallet.getNetwork(),
+      identity: shiplwallet.getIdentity()
     })
 
-    if (shiplwallet.shiplID.auth.identity) {
-      this.pokedex()
+    if (shiplwallet.getIdentity()) {
+      await this.pokedex()
+    } else {
+      this.setState({ textBubble: TextBubble({ text: 'RED, Shipl pay for the gas cost of this transaction. Which POKEMON do you want?' }) })
     }
-  }
-
-  componentDidMount () {
     ReactGA.initialize('UA-136074649-4')
     ReactGA.pageview(window.location.pathname + window.location.search)
   }
 
   async getAccount () {
     const account = (await this.state.web3.eth.getAccounts())[0]
-    this.setState({ identity: this.state.shiplwallet.shiplID.auth.identity })
+    this.setState({ identity: this.state.shiplwallet.getIdentity() })
     return account
   }
 
@@ -237,7 +97,29 @@ class App extends React.Component {
       .claim(pokeId)
       .send({ from: account })
       .on('transactionHash', txHash => this.setState({ txHash }))
-      .on('confirmation', () => this.pokedex())
+      .on('confirmation', (confirmation) => {
+        if (confirmation === 1) this.pokedex()
+      })
+      .on('error', error => {
+        console.error(error)
+      })
+    this.setState({
+      isLoading: true,
+      catch: cards[pokeId - 1].name,
+      txHash: undefined
+    })
+  }
+
+  async claimEvolution (pokeId, pokeIdEvolution) {
+    const TX_VALUE = '1000000000000000000'
+    const account = await this.getAccount()
+    this.state.contract.methods
+      .evolvePokemon(pokeIdEvolution, TX_VALUE)
+      .send({ from: account, value: TX_VALUE })
+      .on('transactionHash', txHash => this.setState({ txHash }))
+      .on('confirmation', (confirmation) => {
+        if (confirmation === 1) this.pokedex()
+      })
     this.setState({
       isLoading: true,
       catch: cards[pokeId - 1].name,
@@ -250,8 +132,20 @@ class App extends React.Component {
     const pokeId = await this.state.contract.methods
       .owners(account)
       .call({ from: account })
-    const poke = cards[pokeId.toNumber() - 1]
-    this.setState({ poke, isLoading: false })
+    const pokeIdNumber = pokeId.toNumber()
+    let cardIndex = pokeIdNumber - 1
+    let poke = cards[cardIndex]
+    if (pokeIdNumber === 11 || pokeIdNumber === 22 || pokeIdNumber === 33) {
+      cardIndex = Math.round(pokeIdNumber / 10) - 1
+      poke = cards[cardIndex]
+      this.setState({ poke, isLoading: false, ownAPokemon: true, textBubble: null, ownAEvolution: true })
+      this.setState({ textBubble: TextBubble({ text: 'RED you already have a POKEMON Evolution! Try other ones!' }) })
+    } else if (pokeIdNumber === 1 || pokeIdNumber === 2 || pokeIdNumber === 3) {
+      this.setState({ poke, isLoading: false, ownAPokemon: true, textBubble: null, ownAEvolution: false })
+      this.setState({ textBubble: TextBubble({ text: 'RED, you can now for 1$ evolve your POKEMON.' }) })
+    } else {
+      this.setState({ textBubble: TextBubble({ text: 'RED, Shipl sponsor you the gas cost of this transaction. Which POKEMON do you want?' }) })
+    }
   }
 
   render () {
@@ -260,17 +154,13 @@ class App extends React.Component {
         <header>
           <div className='container'>
             <div className='nav-brand'>
-              <a
-                href='https://shipl.co'
-                target='_blank'
-                rel='noopener noreferrer'
-              >
+              <a href='https://shipl.co' target='_blank' rel='noopener noreferrer'>
                 <h1>
                   <i className='nes-logo brand-logo' />
                   Poké Shipl
                 </h1>
               </a>
-              <p>A new user experience over the blockchain.</p>
+              <p>A blockchain-less experience.</p>
             </div>
           </div>
         </header>
@@ -285,6 +175,7 @@ class App extends React.Component {
               </p>
             </div>
             <Pokedex
+              ownAEvolution={this.state.ownAEvolution}
               network={this.state.network}
               catch={this.state.catch}
               pokemon={this.state.poke}
@@ -302,26 +193,35 @@ class App extends React.Component {
                   <img
                     className='nes-prof'
                     src='https://vignette.wikia.nocookie.net/pokemontowerdefensetwo/images/c/cd/Professoroak_icon.png/revision/latest?cb=20130710043109'
-                    alt=''
+                    alt='Nes professor'
                   />
                   <div className='nes-balloon from-left'>
-                    <p>Now, RED, which POKEMON do you want?</p>
+                    {this.state.textBubble}
                   </div>
                 </section>
               </section>
-              <div className='nes-container is-rounded'>
-                <div className='card-board'>
-                  {cards.map(e => (
-                    <Card key={e.name} item={e} claim={this.claimPokemon} />
-                  ))}
-                </div>
-              </div>
+              {this.state.web3 &&
+                <div className='nes-container is-rounded'>
+                  {this.state.ownAPokemon && !this.state.ownAEvolution &&
+                    <CardEvolution
+                      claimEvolution={this.claimEvolution}
+                      pictureEvolution={this.state.poke.pictureEvolution}
+                      id={this.state.poke.id}
+                      idEvolution={this.state.poke.idEvolution}
+                      disabled={this.state.web3 === undefined}
+                    />}
+                  <div className='card-board'>
+                    {cards.map(e => (
+                      <Card key={e.name} item={e} claim={this.claimPokemon} disabled={this.state.web3 === undefined} />
+                    ))}
+                  </div>
+                </div>}
             </div>
           </section>
         </div>
         <footer>
           <p>
-            <span>©2019 </span>
+            <span>©2020 </span>
             <a
               href='https://shipl.co'
               target='_blank'
